+++ .gitignore
@@ -1,1 +1,1 @@
-dlbot
+tiktok/tiktok
