+++ common/go.mod
@@ -1,0 +1,5 @@
+module nulo.in/dlbot/common
+
+go 1.19
+
+require github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1 // indirect
