+++ tiktok/go.mod
@@ -2,4 +2,8 @@ module nulo.in/dlbot/tiktok/v4
 
 go 1.18
 
-require github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1 // indirect
+require github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1
+
+require nulo.in/dlbot/common v0.0.0-00010101000000-000000000000 // indirect
+
+replace nulo.in/dlbot/common => ../common
